requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
markdown==3.3.6
django-filter==2.4.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
